backend/pyproject.toml
[tool.poetry]
name = "toolbox-backend"
version = "0.1.0"
description = ""
authors = ["yenslife <77geo5rge6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
